вынес создание фейкового стора
