+++ src/components/movie-card/movie-card.test.js
@@ -3,48 +3,21 @@ import {render, screen} from '@testing-library/react';
 import {Router} from 'react-router-dom';
 import * as redux from 'react-redux';
 import {createMemoryHistory} from 'history';
-import {movieStructure, authInfoStructure} from '../../data-structure';
+import {movieStructure, getFakeStore} from '../../data-structure';
 import {RoutePaths} from '../../utils/constatns';
 import MovieCard from './movie-card';
-import {createAPI} from '../../api/api';
-import rootReducer from '../../store/root-reducer';
-import {configureStore} from '@reduxjs/toolkit';
-import {AuthorizationStatus} from '../../utils/constatns';
-
-const api = createAPI({});
-const mockStore = configureStore({
-  reducer: rootReducer,
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware({
-      thunk: {
-        extraArgument: api
-      },
-    }),
-  preloadedState: {
-    DATA: {
-      selectedMovie: movieStructure,
-      movies: [movieStructure]
-    },
-    USER: {
-      isAuthtorized: AuthorizationStatus.AUTHORIZED,
-      user: authInfoStructure
-    },
-    PLAYER: {
-      movieToPlay: movieStructure,
-      isLoading: true
-    }
-  }
-});
-
-window.HTMLMediaElement.prototype.play = () => {};
-window.HTMLMediaElement.prototype.pause = () => {};
 
 it(`Should MovieCard render correctly`, () => {
+  const fakeStore = getFakeStore();
+
   const history = createMemoryHistory();
   history.push(`${RoutePaths.MY_LIST}`);
 
+  window.HTMLMediaElement.prototype.play = () => {};
+  window.HTMLMediaElement.prototype.pause = () => {};
+
   const {container} = render(
-      <redux.Provider store={mockStore}>
+      <redux.Provider store={fakeStore}>
         <Router history={history}>
           <MovieCard
             movieId={movieStructure.id}
