+++ src/components/my-list/my-list.test.js
@@ -4,41 +4,18 @@ import {Router, Switch} from 'react-router-dom';
 import PrivateRoute from '../../components/private-route/private-route';
 import * as redux from 'react-redux';
 import {createMemoryHistory} from 'history';
-import {movieStructure, authInfoStructure} from '../../data-structure';
+import {getFakeStore} from '../../data-structure';
 import {RoutePaths} from '../../utils/constatns';
 import MyList from './my-list';
-import {createAPI} from '../../api/api';
-import rootReducer from '../../store/root-reducer';
-import {configureStore} from '@reduxjs/toolkit';
-import {AuthorizationStatus} from '../../utils/constatns';
-
-const api = createAPI({});
-const mockStore = configureStore({
-  reducer: rootReducer,
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware({
-      thunk: {
-        extraArgument: api
-      },
-    }),
-  preloadedState: {
-    DATA: {
-      selectedMovie: movieStructure,
-      movies: [movieStructure]
-    },
-    USER: {
-      isAuthtorized: AuthorizationStatus.AUTHORIZED,
-      user: authInfoStructure
-    }
-  }
-});
 
 it(`Should MyList render correctly`, () => {
+  const fakeStore = getFakeStore();
+
   const history = createMemoryHistory();
   history.push(`${RoutePaths.MY_LIST}`);
 
   render(
-      <redux.Provider store={mockStore}>
+      <redux.Provider store={fakeStore}>
         <Router history={history}>
           <Switch>
             <PrivateRoute
